// front/src/app/api-client/model/agenda-entry-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FamilyMemberDTO } from './family-member-dto';


/**
 * Occurrence d\'un reminder dans l\'agenda
 */
export interface AgendaEntryDTO { 
    /**
     * Identifiant de l\'entrée (à utiliser pour le PATCH de statut)
     */
    id?: number;
    reminderId?: number;
    titre?: string;
    description?: string;
    type?: AgendaEntryDTO.TypeEnum;
    /**
     * Date/heure de cette occurrence
     */
    dateHeure?: string;
    /**
     * Fin de cette occurrence (null si le reminder n\'a pas de fin)
     */
    dateHeureFin?: string;
    statut?: AgendaEntryDTO.StatutEnum;
    /**
     * Membres concernés
     */
    membres?: Array<FamilyMemberDTO>;
    /**
     * L\'entrée provient d\'un reminder récurrent
     */
    recurring?: boolean;
}
export namespace AgendaEntryDTO {
    export const TypeEnum = {
        Sport: 'SPORT',
        TestMedical: 'TEST_MEDICAL',
        RendezVous: 'RENDEZ_VOUS',
        Autre: 'AUTRE',
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const StatutEnum = {
        Prevu: 'PREVU',
        Complete: 'COMPLETE',
        Annule: 'ANNULE',
        Deplace: 'DEPLACE',
    } as const;
    export type StatutEnum = typeof StatutEnum[keyof typeof StatutEnum];
}